Move IterableDataProvider to datagen
